DevOps: Move tool migration to `pyproject.toml`
The `mypy.ini` and `tox.ini` configuration files are moved into the
`pyproject.toml` file. The content of the `conftest.py` file in the
repository root is merged with that of `tests/conftest.py`.
