--- mypy.ini
@@ -1,16 +1,0 @@
-[mypy]
-show_error_codes = True
-disallow_untyped_defs = True
-disallow_incomplete_defs = True
-warn_unused_ignores = True
-warn_redundant_casts = True
-no_implicit_optional = True
-; disallow_any_generics = True
-strict_equality = True
-plugins = pydantic.mypy
-
-[pydantic-mypy]
-init_forbid_extra = True
-init_typed = True
-warn_required_dynamic_aliases = True
-warn_untyped_fields = False
