+++ pyproject.toml
@@ -124,3 +124,65 @@ filterwarnings = [
     'ignore::DeprecationWarning:django:',
     'ignore::DeprecationWarning:yaml:',
 ]
+
+[tool.mypy]
+show_error_codes = true
+disallow_untyped_defs = true
+disallow_incomplete_defs = true
+warn_unused_ignores = true
+warn_redundant_casts = true
+no_implicit_optional = true
+strict_equality = true
+plugins = ['pydantic.mypy']
+
+[[tool.pydantic.mypy.overrides]]
+init_forbid_extra = true
+init_typed = true
+warn_required_dynamic_aliases = true
+warn_untyped_fields = false
+
+[tool.tox]
+legacy_tox_ini = """
+[tox]
+envlist = py38
+
+[testenv]
+usedevelop = true
+
+[testenv:py{38,39,310}]
+extras =
+    auth
+    testing
+    docs
+deps = black
+commands = pytest {posargs}
+
+# to get started with doing live testing
+# first create a database folder {toxinidir}/.tox/.aiida/database
+# and run `initdb` in it, then start the postgres server: pg_ctl -D database -l logfile start
+# now you can run `tox -e verdi quicksetup`, then `tox -e serve`
+
+[testenv:verdi]
+setenv =
+    AIIDA_PATH = {toxinidir}/.tox/.aiida
+commands = verdi {posargs}
+
+[testenv:serve]
+extras =
+    auth
+setenv =
+    AIIDA_PATH = {toxinidir}/.tox/.aiida
+commands = uvicorn aiida_restapi:app {posargs:--reload}
+
+[testenv:docs-{update,clean}]
+extras =
+    auth
+    docs
+whitelist_externals =
+    rm
+    echo
+commands =
+    clean: rm -rf docs/_build
+    sphinx-build -nW --keep-going -b {posargs:html} docs/source/ docs/_build/{posargs:html}
+commands_post = echo "open file://{toxinidir}/docs/_build/{posargs:html}/index.html"
+"""
