--- tox.ini
@@ -1,49 +1,0 @@
-# To use tox, see https://tox.readthedocs.io
-# Simply pip or conda install tox
-# If you use conda, you may also want to install tox-conda
-# then run `tox` or `tox -- {pytest args}`
-# run `tox -r` to recreate the environment
-# run e.g. `tox -e serve` to select a particular environment
-
-[tox]
-envlist = py38
-
-[testenv]
-usedevelop = true
-
-[testenv:py{38,39,310}]
-extras =
-    auth
-    testing
-    docs
-deps = black
-commands = pytest {posargs}
-
-# to get started with doing live testing
-# first create a database folder {toxinidir}/.tox/.aiida/database
-# and run `initdb` in it, then start the postgres server: pg_ctl -D database -l logfile start
-# now you can run `tox -e verdi quicksetup`, then `tox -e serve`
-
-[testenv:verdi]
-setenv =
-    AIIDA_PATH = {toxinidir}/.tox/.aiida
-commands = verdi {posargs}
-
-[testenv:serve]
-extras =
-    auth
-setenv =
-    AIIDA_PATH = {toxinidir}/.tox/.aiida
-commands = uvicorn aiida_restapi:app {posargs:--reload}
-
-[testenv:docs-{update,clean}]
-extras =
-    auth
-    docs
-whitelist_externals =
-    rm
-    echo
-commands =
-    clean: rm -rf docs/_build
-    sphinx-build -nW --keep-going -b {posargs:html} docs/source/ docs/_build/{posargs:html}
-commands_post = echo "open file://{toxinidir}/docs/_build/{posargs:html}/index.html"
